fixed c# acces denied problems

+++ csharp/FileSystem.cs
@@ -41,13 +41,14 @@ namespace Cs
 
             try
             {
-                if (File.Exists(str_path))
-                    File.Delete(str_path);
+                if (File.Exists(str_path)) File.Delete(str_path);
                 
                 using (StreamWriter sw = File.CreateText(str_path))
                 {
                     sw.WriteLine(str_content);
+                    sw.Close();
                 }
+
             }
             catch (Exception e)
             {
